Review fixes for user_service.py
